alteracao da camada de modelo e criacao dos arquivo hbm.xml

+++ card/card-model/card-model-domain/src/main/java/br/com/bancoamazonia/integracao/card/model/domain/Evento.java
@@ -6,6 +6,6 @@ package br.com.bancoamazonia.integracao.card.model.domain;
  */
 public class Evento {
 	//id
-	private Integer codigo;
+	private Byte codigo;
 	private String descricao;
 }
