+++ card/card-model/card-model-domain/src/main/java/br/com/bancoamazonia/integracao/card/model/domain/Limite.java
@@ -3,7 +3,6 @@ package br.com.bancoamazonia.integracao.card.model.domain;
 import java.math.BigDecimal;
 
 public class Limite {
-
-	private Long id;
+	private Byte id;
 	private BigDecimal valor;
 }
